requirements/pypi.txt
# Django
django

# Social Login
python-social-auth
social-auth-app-django

# OAuth2 Provider
django-oauth-toolkit
django-cors-headers

# REST 
djangorestframework
